requirements.txt
requests
beautifulsoup4
python-telegram-bot==13.15
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psutil==5.9.6
ray[rllib]==2.8.0
ray[tune]==2.8.0
ray[serve]==2.8.0
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.36.0
accelerate==0.25.0
datasets==2.15.0
wandb==0.16.1
gymnasium==0.29.1
stable-baselines3==2.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
ipywidgets==8.1.1
